requirements.txt
cdifflib
argparse
termcolor
asyncio==3.4.3
aiohttp==3.6.2
docker==4.0.2
python-libnmap==0.7.0
requests
tqdm==4.37.0
selenium==3.141.0
string;python_full_version<"3.7.4"
random;python_full_version<"3.7.4"
